frontend: add react-router-dom, AppShell/TopNav, setup gate
Restructures App.jsx into router + shell + setup gate (redirects to
/onboarding until GET /setup/state reports github_connected,
repos_selected, and first_index_done). Existing chat UI moves to
ask/AskPage.jsx unchanged; Library/Settings routes get placeholder
content until their own issues land.
