+++ frontend/src/ask/AskPage.jsx
@@ -1,0 +1,126 @@
+import { useEffect, useState } from 'react'
+import { getRepos, postQuery } from '../api.js'
+import ChatInput from '../components/ChatInput.jsx'
+import MessageList from '../components/MessageList.jsx'
+import RepoFilter from '../components/RepoFilter.jsx'
+
+const EXAMPLE_QUESTIONS = [
+  'Why is authentication done this way?',
+  'What alternatives did we consider for the database?',
+  'Who made the decision to use Redis, and when?',
+]
+
+function AskPage() {
+  const [repos, setRepos] = useState(undefined)
+  const [selectedRepos, setSelectedRepos] = useState([])
+  const [messages, setMessages] = useState([])
+  const [loading, setLoading] = useState(false)
+  const [chatKey, setChatKey] = useState(0)
+  const [prefill, setPrefill] = useState('')
+
+  useEffect(() => {
+    let cancelled = false
+
+    getRepos()
+      .then((result) => {
+        if (cancelled) return
+        setRepos(result.repos)
+        setSelectedRepos(result.repos.map((repo) => repo.repo))
+      })
+      .catch(() => {
+        if (cancelled) return
+        setRepos('error')
+      })
+
+    return () => {
+      cancelled = true
+    }
+  }, [])
+
+  function replaceLastMessage(message) {
+    setMessages((prev) => [...prev.slice(0, -1), message])
+  }
+
+  async function runQuery(question) {
+    setLoading(true)
+    try {
+      const result = await postQuery({ question, repos: selectedRepos })
+      replaceLastMessage({
+        role: 'assistant',
+        type: 'answer',
+        answer: result.answer,
+        citations: result.citations,
+      })
+    } catch (err) {
+      replaceLastMessage({
+        role: 'assistant',
+        type: 'error',
+        message: err.message,
+        onRetry: () => retry(question),
+      })
+    } finally {
+      setLoading(false)
+    }
+  }
+
+  function retry(question) {
+    replaceLastMessage({ role: 'assistant', type: 'loading' })
+    runQuery(question)
+  }
+
+  function handleAsk(question) {
+    setMessages((prev) => [
+      ...prev,
+      { role: 'user', content: question },
+      { role: 'assistant', type: 'loading' },
+    ])
+    runQuery(question)
+  }
+
+  function handleChipClick(question) {
+    setPrefill(question)
+    setChatKey((key) => key + 1)
+  }
+
+  return (
+    <div className="h-full flex flex-col">
+      <header className="h-14 shrink-0 bg-panel flex items-center justify-between px-4">
+        <RepoFilter repos={repos} selected={selectedRepos} onChange={setSelectedRepos} />
+      </header>
+
+      <main className="flex-1 overflow-y-auto px-4 py-4">
+        {messages.length === 0 ? (
+          <div className="h-full flex flex-col items-center justify-center gap-4 text-center">
+            <p className="text-base text-ink-muted">
+              Ask why something in your codebase is the way it is
+            </p>
+            <div className="flex flex-wrap justify-center gap-4">
+              {EXAMPLE_QUESTIONS.map((question) => (
+                <button
+                  key={question}
+                  type="button"
+                  onClick={() => handleChipClick(question)}
+                  className="rounded-lg border border-transparent bg-panel text-ink text-sm px-4 py-2 hover:border-accent"
+                >
+                  {question}
+                </button>
+              ))}
+            </div>
+          </div>
+        ) : (
+          <div className="max-w-3xl mx-auto">
+            <MessageList messages={messages} disabled={loading} />
+          </div>
+        )}
+      </main>
+
+      <div className="sticky bottom-0 shrink-0 bg-panel px-4 py-4">
+        <div className="max-w-3xl mx-auto">
+          <ChatInput key={chatKey} initialValue={prefill} onSubmit={handleAsk} disabled={loading} />
+        </div>
+      </div>
+    </div>
+  )
+}
+
+export default AskPage
